Simplify Tiltify setup
