+++ src/extension/index.ts
@@ -1,32 +1,22 @@
-// Move this into src, then have parcel generate extensions. MUCH easier, can also moves types
-
 import * as path from "path";
-import { CampaignClient } from "./clients/tiltify-client";
+import setupTiltify from "./clients/tiltify-client";
 import { TwitchClient } from "./clients/twitch-client";
-import milestones from "./domains/milestones";
-import donations from "./domains/donations";
-import campaign from "./domains/campaign";
 import bot from "./bot";
 // import secrets from "./secrets";
 
 import type { NodeCG } from "nodecg-types/types/server";
-import type { CampaignClientArgs } from "./clients/tiltify-client";
 
 import * as TILTIFY_CONFIG from "../../config/tiltify.json";
 
 // Inject any dependencies manually.
 export default async (nodecg: NodeCG) => {
   // Set up dependencies
-  const campaignClient = new CampaignClient(
-    TILTIFY_CONFIG as CampaignClientArgs
-  );
   const twitchClient = await TwitchClient.create(
     path.join(__dirname, "../../config/twitch.json")
   );
 
-  milestones(nodecg, campaignClient);
-  donations(nodecg, campaignClient);
-  campaign(nodecg, campaignClient);
+  setupTiltify(nodecg, TILTIFY_CONFIG);
+
   bot(nodecg, twitchClient);
   // secrets(nodecg, twitchClient);
 };
